+++ RevUp-infra/src/main/java/com/revup/redis/config/RedissonConfig.java
@@ -1,4 +1,4 @@
-package com.revup.feedback;
+package com.revup.redis.config;
 
 import org.redisson.Redisson;
 import org.redisson.api.RedissonClient;
